Fix #3
PayloadTypeIdentifier's constructor is modified to accept an optional
delimiter, which is a sequence of characters that identifies the start
of a payload, if any.

The delimiter defaults to a double CRLF if no value is specified.

The functionality of Utils.IndexOfPayload() is moved to
PayloadTypeIdentifier.IndexOfPayload(), which is the one that is now
called by RequestRecord and ResponseRecord.

Users who need to parse content blocks that have a payload with a
custom delimiter should override PayloadTypeIdentifier and pass it to a
RecordFactory as follows:

new RecordFactory(payloadTypeIdentifier: new CustomPayloadTypeIdentifier());