Fix #5
In parse mode – i.e. when using the `WarcParser` - any subclass of `Record` that has a `WARC-Identified-Payload-Type` will be created with its `IdentifiedPayloadType` set to the parsed value, if any.

In non parse mode – i.e. when instantiating each of those subclass using their respective constructor(s) – the ` IdentifiedPayloadType` will be set to the value of `PayloadTypeIdentifier.Identify(…)`.

This bug is due to an oversight as it should have operated the same way as `Record.BlockDigest`, which implemented the conditional check from the get-go.